NABI-304--refactor : 단건 응답 패키지 수정
